refactor(search): moved search/search into just search
