+++ src/search/engines/name.go
@@ -5,7 +5,7 @@ import "strings"
 type Name uint8
 
 //go:generate enumer -type=Name -json -text -yaml -sql
-//go:generate go run github.com/hearchco/hearchco/generate/searcher -type=Name -packagename search -output ../search/engine_searcher.go
+//go:generate go run github.com/hearchco/hearchco/generate/searcher -type=Name -packagename search -output ../engine_searcher.go
 const (
 	UNDEFINED Name = iota
 	BING
